% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent2edge.R
\name{adjacent2edge}
\alias{adjacent2edge}
\title{Function to convert tree in adjacent list format to edge
representation.}
\usage{
adjacent2edge(tree)
}
\arguments{
\item{tree}{A data.table object with two columns.  The first column should
have the parent nodes and the second column should have a character vector
containing all the children nodes separated by commas (i.e. adjacent list
format).}
}
\value{
A data.table object containing two columns: parent and children.
}
\description{
Function to convert tree in adjacent list format to edge
representation.
}
\examples{

library(data.table)
children <- matrix(sample(letters, 4 * 3, replace = TRUE), nrow=3)
children <- apply(children, 1, paste0, collapse = ", ")
adj <- data.table(parent = c("a", "b", "c"), children = children)

adjacent2edge(adj)

}

