% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkMethodFlag.R
\name{checkMethodFlag}
\alias{checkMethodFlag}
\title{Check method flag}
\usage{
checkMethodFlag(flag)
}
\arguments{
\item{flag}{A character (or character vector) of flag values to test for
validity.}
}
\value{
A logical vector of the same length as flag indicating if the flags
are valid (TRUE = valid).
}
\description{
Method flags in the SWS are only allowed to take certain values.  This
function ensures that the supplied flags are all valid.
}

