% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProtectedData.R
\name{checkProtectedData}
\alias{checkProtectedData}
\title{Function to check whether the data will over write protected data.}
\usage{
checkProtectedData(dataToBeSaved, domain = "agriculture",
  dataset = "aproduction", areaVar = "geographicAreaM49",
  itemVar = "measuredItemCPC", elementVar = "measuredElement",
  yearVar = "timePointYears", flagObservationVar = "flagObservationStatus",
  flagMethodVar = "flagMethod", protectedFlag = c("-", "q", "p", "h", "c"))
}
\arguments{
\item{dataToBeSaved}{The data.table object containing data to be
saved back to the database. The current implementation only
accepts data in the normalised form.}

\item{domain}{The domain name in the SWS where the data will be
saved.}

\item{dataset}{The dataset name in the SWS where the data will be
saved.}

\item{areaVar}{The column name corresponding to the geographic
area.}

\item{itemVar}{The column name corresponding to the commodity
item.}

\item{elementVar}{The column name corresponding to the measured
element.}

\item{yearVar}{The column name corresponding to the year.}

\item{flagMethodVar}{The column name corresponding to the method flag.}

\item{protectedFlag}{The set of flag values which corresponds to
values that should not be over written.}
}
\value{
If the data set passes the test, then the original data
    will be returned. Otherwise an error will be raised.
}
\description{
Certain data in the Statistical Working System should not be over
written by algorithms. Namely, official and semi-official
values. This function takes the data to be saved in the arguement
and pulls the exact same set of data from the data base, then
checks whether the matching set contains official or semi official
values.
}

