% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTimeSeriesImputed.R
\name{checkTimeSeriesImputed}
\alias{checkTimeSeriesImputed}
\title{This function checks whether all time series within a dataset are imputed.}
\usage{
checkTimeSeriesImputed(dataToBeSaved, key, valueColumn)
}
\arguments{
\item{dataToBeSaved}{The data to be saved back to the database}

\item{key}{The key which splits the data into individual timeseries.
Generally the set c('geographicAreaM49', 'measuredItemCPC',
'measuredElement').}

\item{valueColumn}{The column which contains the numeric values.}
}
\value{
The same data if all time series are imputed, otherwise an error.
}
\description{
This function checks whether all time series within a dataset are imputed.
}

