% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeYield.R
\name{computeYield}
\alias{computeYield}
\title{Function to compute and update yield}
\usage{
computeYield(data, productionValue, productionObservationFlag,
  areaHarvestedValue, areaHarvestedObservationFlag, yieldValue,
  yieldObservationFlag, yieldMethodFlag, newObservationFlag = "I",
  newMethodFlag = "i", flagTable, unitConversion = 1)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{productionValue}{The column name corresponding to production value.}

\item{productionObservationFlag}{The column name corresponding to the
observation flag of production.}

\item{areaHarvestedValue}{The column name corresponding to area harvested
value.}

\item{areaHarvestedObservationFlag}{The column name corresponding to the
observation flag of area harvested.}

\item{yieldValue}{The columne name corresponding to yield value.}

\item{yieldObservationFlag}{The column name corresponding to the observation
flag of yield.}

\item{yieldMethodFlag}{The column name corresponding to the method flag for
yield.}

\item{newObservationFlag}{The character value that should be placed in the
yieldMethodFlag column when yield is computed.}

\item{newMethodFlag}{The character value that should be placed in the
yieldMethodFlag column when yield is computed.}

\item{unitConversion}{yield is computed as (production)/(area harvested)*
  (unit conversion).}

\item{Currently}{not used.  Was originally implemented for flag aggregation,
but the decision was to alway have yield written as "I"/"i".}
}
\description{
Function to compute and update yield
}

