% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/containInfo.R
\name{containInfo}
\alias{containInfo}
\title{Function to determine whether the data contains any information.}
\usage{
containInfo(value, observationFlag, naFlag = "M")
}
\arguments{
\item{value}{A numeric vector to be checked.}

\item{observationFlag}{The character vector representing the observation
flag corresponding to the value.}

\item{naFlag}{The value of the observation flag which indicates that data
is missing.  Defaults to "M".}
}
\value{
A logical value (TRUE or FALSE) indicating whether or not the
passed vectors contain any information.
}
\description{
If the data contains only missing values or zeroes then it is
marked as containing no information.
}

