% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryExists.R
\name{countryExists}
\alias{countryExists}
\title{Find if country exists for a given year}
\usage{
countryExists(countries, years, names = FALSE, warnNA = TRUE)
}
\arguments{
\item{countries}{vector of countries. Can be either names of M49 codes}

\item{years}{vector of years, must be same length as countries}

\item{names}{logical. TRUE when names are given. They must correspond to the 
descriptions in agriculture:agriculture}

\item{warnNA}{logical. By default, a warning will be given when an NA is in 
countries or year. Setting this to \code{FALSE} suppresses it.}
}
\value{
A vector (TRUE/FALSE) of the same length as countries and year 
  corresponding to if they are a country or not. A warning - if this has NAs,
  subsetting by an NA can be problematic, i.e. result in no matches when
  subsetting.
}
\description{
Some countries like the former Soviet union or South Sudan dissolve or are 
formed during the period where we collect data. This function when given a 
country or a year returns a boolean expressing if the country exists for that
year.
}

