% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpc2fcl.R
\name{cpc2fcl}
\alias{cpc2fcl}
\title{CPC to FCL Codes}
\usage{
cpc2fcl(cpcCodes)
}
\arguments{
\item{cpcCodes}{A character vector containing the CPC codes.}
}
\value{
A character vector whose ith element is the FCL code corresponding
  to the passed CPC code in position i.  If no valid mapping is found, an NA
  is returned in this place.
}
\description{
This function maps CPC to FCL codes using the table defined within the SWS.
}
\examples{
\dontrun{
cpc2fcl("0111", "23110", "01447")
cpc2fcl(fcl2cpc(c("0015", "0016", "0265")))
}
}

