% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denormalise.R
\name{denormalise}
\alias{denormalise}
\title{This function denormalise data in the SWS format}
\usage{
denormalise(normalisedData, denormaliseKey, areaVar = "geographicAreaM49",
  itemVar = "measuredItemCPC", elementVar = "measuredElement",
  yearVar = "timePointYears", flagObsVar = "flagObservationStatus",
  flagMethodVar = "flagMethod", valueVar = "Value",
  fillEmptyRecords = FALSE)
}
\arguments{
\item{normalisedData}{normalised data to be denormalise}

\item{areaVar}{The column name corresponding to the geographic area.}

\item{itemVar}{The column name corresponding to the commodity item.}

\item{elementVar}{The column name corresponds to the measured element.}

\item{yearVar}{The column name corresponds to the time dimension.}

\item{flagObsVar}{The column name corresponds to the observation status
flag.}

\item{flagMethodVar}{The column name corresponds to the method flag.}

\item{valueVar}{The column name corresponds to the value columne.}

\item{fillEmptyRecords}{logical, whether empty record created due to
denormalisation should be replaced with missing value.}

\item{denormalisedKey}{The key/variable that will be denormalised.}
}
\value{
The denormalised data
}
\description{
This function denormalise data in the SWS format
}

