% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{faoswsEUtil-package}
\alias{faoswsEUtil-package}
\title{A suite of function to support the manipulation of data for
working with the FAO new statistical working system.}
\description{
This package provides extra functionality to the faosws package
for the manipulation of data.
}
\author{
Michael. C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}

