% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcl2cpc.R
\name{fcl2cpc}
\alias{fcl2cpc}
\title{FCL to CPC Codes}
\usage{
fcl2cpc(fclCodes, version = "2.1preliminary")
}
\arguments{
\item{fclCodes}{A character vector containing the FCL codes.}

\item{version}{character. This specifies the version of the conversion 
tables between fcl and cpc. The only ones available, so far, are '2.1preliminary',
the default one, and '2.1' which is the update and official one. Ideally in 
the future the default value should be '2.1'.}
}
\value{
A character vector whose ith element is the CPC code corresponding 
  to the passed FCL code in position i.  If no valid mapping is found, an NA
  is returned in this place.
}
\description{
This function maps FCL to CPC codes using the table defined within the SWS.
}
\examples{
\dontrun{
fcl2cpc(c("0015", "0016", "0265"))
fcl2cpc(cpc2fcl(c("0111", "23110", "01447")))
}

}

