% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillRecord.R
\name{fillRecord}
\alias{fillRecord}
\title{This function replaces non-existing records with missing records.}
\usage{
fillRecord(data, areaVar = "geographicAreaM49", itemVar = "measuredItemCPC",
  elementVar = "measuredElement", yearVar = "timePointYears",
  valueVar = "Value", flagObsVar = "flagObservationStatus",
  flagMethodVar = "flagMethod")
}
\arguments{
\item{data}{The dataset}

\item{areaVar}{The column name corresponding to the geographic area.}

\item{itemVar}{The column name corresponding to the commodity item.}

\item{elementVar}{The column name corresponds to the measured element.}

\item{yearVar}{The column name corresponds to the time dimension.}

\item{valueVar}{The regular expression to capture the value column.}

\item{flagObsVar}{The regular expression to capture the observation
status flag column.}

\item{flagMethodVar}{The regular expression to capture the method flag
column.}
}
\value{
dataset with identical dimension but creates missing record in place
    of non-existing record.
}
\description{
Occastionally the database may return records where all three columns
(value, observation flag, method flag) are all NA and is a non-existing
record in the database; this can also be created when denormalising data. We
will replace is with a missing record representation where (Value = NA,
flagObservationStatus = "M" and flagMethod = "u").
}

