% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs2m49.R
\name{fs2m49}
\alias{fs2m49}
\title{FS to M49 Codes}
\usage{
fs2m49(fsCodes)
}
\arguments{
\item{fsCodes}{A character vector containing the FS codes.}
}
\value{
A character vector whose ith element is the M49 code corresponding 
  to the passed FS code in position i.  If no valid mapping is found, an NA
  is returned in this place.
}
\description{
This function maps FS to M49 (country) codes using the table defined within
the SWS.
}
\examples{
\dontrun{
fs2m49(c("1", "2", "3", "4"))
m492fs(fs2m49(c("1", "2", "3", "4")))
}

}
