% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllData.R
\name{getAllData}
\alias{getAllData}
\title{Get all data from a table in the Statistical Working System}
\usage{
getAllData(domain, dataset)
}
\arguments{
\item{domain}{character. Name of the domain}

\item{dataset}{character. Name of the dataset}
}
\description{
In some rare cases, you may want to fetch all data from a Statistical Working
System table. This function uses the code list from the dataset and makes a 
key will all codes from all dimensions. It then gets that data. For obvious 
reasons, this can take some time. The domain and dataset name can be
retrieved from the \code{swsContext.datasets} or you can use the
\code{\link{GetDomainNames}} family of functions.
}
