% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChildren.R
\name{getChildren}
\alias{getChildren}
\title{Get Children}
\usage{
getChildren(commodityTree, parentColname, childColname, topNodes,
  parentFlag = TRUE)
}
\arguments{
\item{commodityTree}{A data.table with parent and child node IDs 
(corresponding to the IDs in nodes) which specify the commodity tree 
structure.}

\item{parentColname}{The column name of commodityTree which contains the ID 
of the parent node.}

\item{childColname}{The column name of commodityTree which contains the ID 
of the child node.}

\item{topNodes}{Vector containing the node names/ids for which the children 
are sought.}

\item{parentFlag}{Logical.  If TRUE (default) this function gets all 
descendants.  If FALSE, it gets all ancestors.}
}
\value{
A numeric vector containing topNodes as well as all descendants (or
  ancestors) of topNodes.
}
\description{
This function takes a commodity tree and a vector of node ids and returns a 
vector of these nodes as well as all the descendants (or ancestors) of these
nodes.
}
