% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommodityLevel.R
\name{getCommodityLevel}
\alias{getCommodityLevel}
\title{Get Commodity Level}
\usage{
getCommodityLevel(commodityTree, parentColname, childColname,
  returnMinLevel = TRUE)
}
\arguments{
\item{commodityTree}{A data.table with parent and child node IDs 
(corresponding to the IDs in nodes) which specify the commodity tree 
structure.}

\item{parentColname}{The column name of commodityTree which contains the ID 
of the parent node.}

\item{childColname}{The column name of commodityTree which contains the ID 
of the child node.}

\item{returnMinLevel}{Logical.  If a node exists at multiple processing 
levels, should the minimum level be returned?  Or, should multiple records
be returned for each level?}
}
\value{
A data.table with two columns: node (the ID of the commodity code) 
  and level.  A level of 0 indicates a top level node which is then 
  processed into a level 1 node.  Level 1 nodes are processed into level 2, 
  and so on.
}
\description{
This function takes a commodity tree and provides the level of each 
commodity within that tree.
}

