% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompleteImputationKey.R
\name{getCompleteImputationKey}
\alias{getCompleteImputationKey}
\title{This function reads the complete imputation key from the Datatable
'fbs_[table]_comm_codes'}
\usage{
getCompleteImputationKey(table = "production", tableOnly = FALSE)
}
\arguments{
\item{table}{The Datatable to read from.}

\item{tableOnly}{logical, if true only the table is returned. The default is
to return the DatasetKey.}
}
\value{
A DatasetKey object containing all the keys for production
    imputation
}
\description{
This function reads the complete imputation key from the Datatable
'fbs_[table]_comm_codes'
}
