% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNutritiveFactors.R
\name{getNutritiveFactors}
\alias{getNutritiveFactors}
\title{Get Nutritive Factors}
\usage{
getNutritiveFactors(geographicAreaM49 = NULL, measuredElement = NULL,
  measuredItemCPC = NULL, timePointYearsSP = NULL)
}
\arguments{
\item{geographicAreaM49}{A character vector of area codes.  "0" (wildcard) 
will always be included if it is not passed, but passing it does not cause
any issues.  If NULL, the function pulls all area codes from the dimension
table.}

\item{measuredElement}{A character vector of measuredElement codes.  These 
are specific to the nutrient factors table (i.e. Aupus Ratios).  Calories,
proteins, and fats are 1001, 1003, and 1005, respectively.  If NULL, the 
function pulls all area codes from the dimension table.}

\item{measuredItemCPC}{A character vector of CPC codes for the commodities 
of interest.  If NULL, the function pulls all area codes from the 
dimension table.}

\item{timePointYearsSP}{A character vector of years.  As with 
geographicAreaM49, "0" need not be passed but can be.  If NULL, the 
function pulls all area codes from the dimension table.

NOTE: This function chooses the nutrient factor for a particular 
country/year by first looking for data specific to that country/year, then
data specific to that country and all years, then data specific to that 
year and all countries, then data for all years and all countries.  This
order may not be appropriate, but it was my best guess at the time of
writing the code.}
}
\value{
A data.table containing the nutrient factors.  Any nutrient factor 
  that applies to multiple years/countries is now represented as multiple 
  rows in the table.
}
\description{
This function pulls nutrient factor data from the SWS.  The tricky part is 
that some nutrient factors are specific to country/year, and some apply to 
all years or all countries or both.  Thus, we must combine these factors in 
a reasonable way and generate a useable table.
}

