% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShareData.R
\name{getShareData}
\alias{getShareData}
\title{This function obtains the share data.}
\usage{
getShareData(geographicAreaM49, measuredItemChildCPC, measuredItemParentCPC,
  timePointYearsSP)
}
\arguments{
\item{geographicAreaM49}{A character vector of area codes.  The trees
returned are specific to country and year; thus, providing this parameter
limits which trees are pulled.  If NULL, all are used.}

\item{timePointYearsSP}{A character vector of years.  See geographicAreaM49.}
}
\value{
A table of the share
}
\description{
This function obtains the share data.
}

