% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldElements.R
\name{getYieldCodes}
\alias{getYieldCodes}
\title{Get Yield Codes}
\usage{
getYieldCodes(itemCode)
}
\arguments{
\item{itemCode}{A character vector of CPC codes.}
}
\value{
A data.table with the passed CPC codes as well as the corresponding
  element codes for the production, yield, and area harvested elements.  In
  addition, "factor" is returned, and this satisfies the formula
  
Yield = Production / (Area Harvested) * factor
}
\description{
This function takes a vector of CPC codes and returns a data.table with the 
corresponding production/yield/area harvested codes for those commodities.
}

