% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m492fs.R
\name{m492fs}
\alias{m492fs}
\title{M49 to FS Codes}
\usage{
m492fs(m49Codes)
}
\arguments{
\item{m49Codes}{A character vector containing the M49 codes.}
}
\value{
A character vector whose ith element is the FS code corresponding 
  to the passed M49 code in position i.  If no valid mapping is found, an NA
  is returned in this place.
}
\description{
This function maps M49 to FS (country) codes using the table defined within
the SWS.
}
\examples{
\dontrun{
m492fs(c("1", "2", "3", "4"))
m492fs(fs2m49(c("1", "2", "3", "4")))
}

}

