% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEmptyDataset.R
\name{makeEmptyDataset}
\alias{makeEmptyDataset}
\title{Make empty data.table based on SWS table.}
\usage{
makeEmptyDataset(domain, dataset)
}
\arguments{
\item{domain}{character. SWS domain}

\item{dataset}{character. SWS dataset}
}
\description{
Make empty data.table based on SWS table.
}

