% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKeyDt.R
\name{makeKeyDt}
\alias{makeKeyDt}
\title{Convert DatasetKey to data.table}
\usage{
makeKeyDt(dskey)
}
\arguments{
\item{dskey}{DatasetKey. An object that you've either created yourself or
retrieved from the swsContext.datasets[[1]] object}
}
\description{
Keys themselves can be hard to work with and share. This function converts a 
DatasetKey to a table format to make it easier to work with.
}
\author{
Sebastian Campbell <sebastian.campbell@fao.org>
}

