% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na2zero.R
\name{na2zero}
\alias{na2zero}
\title{NA to zero}
\usage{
na2zero(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
The value of x if x has a value and 0 if x is na or has length 0.
  If x has length >1, an error is thrown.
}
\description{
Helper function to convert NA's to 0 when summing.
}

