% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameData.R
\name{nameData}
\alias{nameData}
\title{Name SWS dataset}
\usage{
nameData(domain, dataset, dt, except, append = "_description")
}
\arguments{
\item{domain}{character. Domain of dataset}

\item{dataset}{character. Name of dataset}

\item{dt}{data.table. The data.table to which you wish to add names}

\item{except}{character. Vector of names in the table that you do not wish
named. For example, "timePointYears"}

\item{append}{character. suffix to add to column name of names}
}
\description{
Adds human-readable names to keys automatically.
}
\examples{
\dontrun{
dat = GetData(key)
nameData("agriculture", "aproduction", dat, except = "timePointYears")
}

}

