% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{This function normalise the denormalised data assuming the
structure given by the SWS specification.}
\usage{
normalise(denormalisedData, areaVar = "geographicAreaM49",
  itemVar = "measuredItemCPC", elementVar = "measuredElement",
  yearVar = "timePointYears", flagObsVar = "flagObservationStatus",
  flagMethodVar = "flagMethod", valueVar = "Value")
}
\arguments{
\item{denormalisedData}{The data.table object which is
denormalised in the form returned by the \code{GetData}
function when the normalised arguement is set to FALSE.}

\item{areaVar}{The column name corresponding to the geographic
area.}

\item{itemVar}{The column name corresponding to the commodity
item.}

\item{elementVar}{The column name corresponding to the measured
element.}

\item{yearVar}{The column name corresponding to the year.}

\item{flagObsVar}{The column name corresponding to the observation
status flag.}

\item{flagMethodVar}{The column name corresponding to the method
flag.}

\item{valueVar}{The column name corresponding to the value.}
}
\value{
A normalised data.table object in the same form returned
    by the \code{GetData} function when the normalised arguement
    is set to TRUE.
}
\description{
This function normalise the denormalised data assuming the
structure given by the SWS specification.
}

