% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NullToNa.R
\name{NullToNa}
\alias{NullToNa}
\title{Null to NA}
\usage{
NullToNa(nullList)
}
\arguments{
\item{nullList}{The input list of data which is to be checked for NULLs.}
}
\value{
A vector of the same length as nullList, but with NULL replaced
with NA.
}
\description{
This function takes a vector which may contain NULL values and converts
any NULLs to NAs.
}

