% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDiff.R
\name{packageDiff}
\alias{packageDiff}
\title{Function to compare two sets of packages}
\usage{
packageDiff(source_pkgs, dest_pkgs, exclude_base = c("base", "recommended"))
}
\arguments{
\item{source_pkgs}{data.frame}

\item{dest_pkgs}{data.frame.}

\item{exclude_base}{character. Package priority types to exclude. If you
don't want to exclude any, set this to NULL}
}
\description{
Function to compare two sets of packages
}
\keyword{internal}
