% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRange.R
\name{parseRange}
\alias{parseRange}
\title{Parse range string}
\usage{
parseRange(range, sep = ", *", rsep = " *- *")
}
\arguments{
\item{range}{character. By default, a character vector with dashes to
delineate ranges and commas to separate them}

\item{sep}{character Regular expression to for range separator}

\item{rsep}{character Regular expression for range delineator}
}
\value{
character vector of numbers
}
\description{
As with printer page ranges, sometimes you need to process a range. This is
usually of years. A user may specify that they want 2000, 2002 and 2005-2008.
}
\examples{
parseRange("5, 7-9, 11")
# [1] "5"  "7"  "8"  "9"  "11"

parseRange("2*3;0", sep = "; *", rsep = "\\\\* *")
# [1] "2" "3" "0"

}
