% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateDimension.R
\name{preprocess_key_tree}
\alias{preprocess_key_tree}
\title{Preprocess key tree}
\usage{
preprocess_key_tree(data, dim_name, key_table_in)
}
\arguments{
\item{data}{The data to be aggregated}

\item{dim_name}{The name of the dimension in the data to aggregate}

\item{key_table_in}{The key tree to prepare in the form data.table(parent,children)}
}
\value{
The prepared key tree with an added "level" column
}
\description{
This performs two functions, #1 it removes any key values
not present in the supplied data, #2 it parses the tree and gives each entry
a level number, where level zero is where the parent of the branch has not
parents of its own (i.e. is a root)
}
\keyword{internal}

