% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove0M.R
\name{remove0M}
\alias{remove0M}
\title{This function replaces value which are zero and has flag
representing missing values with NA.}
\usage{
remove0M(data, valueVars, flagVars, missingFlag = "M")
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{valueVars}{The value of the observation.}

\item{flagVars}{The flag of the observation.}

\item{missingFlag}{The value of the flag which denotes missing value.}
}
\value{
The data object with all the (0, M) replaced with (NA, M)
}
\description{
TODO (Michael): This function is similar to the function in the
                \pkg{faoswsUtil} package, but enables multiple
                value and flag pairs and also returning the data
                object. This function should replace the original
                one in the \pkg{faoswsUtil} package
}

