% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeImputation.R
\name{removeImputation}
\alias{removeImputation}
\title{Function to remove prior imputation.}
\usage{
removeImputation(data, value, observationFlag, methodFlag,
  missingObservationFlag = "M", missingMethodFlag = "u",
  imputedFlag = "I")
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{value}{The column name of data which contains the values for the
variable to be modified.}

\item{observationFlag}{The column name of the observation flag
corresponding to value.}

\item{methodFlag}{The column name of the method flag corresponding to
value.}

\item{missingObservationFlag}{The character value which is used to
represent missing values in the observation flag.  This value will be put
in any column containing imputation data that is removed.}

\item{missingMethodFlag}{The character value which is used to represent
missing values in the method flag.  This value will be put in any column
containing imputation data that is removed.}

\item{imputedFlag}{This character value specifies which observation flags
correspond to imputed values in the data.}
}
\value{
No value is returned.  However, the object "data" which was passed
to this function is modified.
}
\description{
Function to remove prior imputation.
}
\details{
Observations which have observationFlag == imputedFlag will be
modified to missing.  Thus, the value will be updated to NA, the
observation flag will be updated to missingObservationFlag, and the method
flag will be updated to missingMethodFlag.
}

