% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeImputationEstimation.R
\name{removeImputationEstimation}
\alias{removeImputationEstimation}
\title{Function to remove prior imputation.}
\usage{
removeImputationEstimation(data, valueVar, observationFlagVar, methodFlagVar,
  missingObservationFlag = "M", missingMethodFlag = "u",
  imputationEstimationObservationFlag = "I",
  imputationEstimationMethodFlag = "e")
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{valueVar}{The column name of data which contains the values for the
variable to be modified.}

\item{observationFlagVar}{The column name of the observation flag
corresponding to value.}

\item{methodFlagVar}{The column name of the method flag corresponding to
value.}

\item{missingObservationFlag}{The character value which is used to represent
missing values in the observation flag. This value will be put in any
column containing imputation data that is removed.}

\item{missingMethodFlag}{The character value which is used to represent
missing values in the method flag. This value will be put in any column
containing imputation data that is removed.}

\item{imputationEstimationObservationFlag}{This character value specifies
which observation flags correspond to imputed/estimated values in the
data.}

\item{imputationEstimationMethodFlag}{This character value specifies which
method flags correspond to imputed/estimated values in the data.}
}
\value{
No value is returned. However, the object "data" which was passed to
    this function is modified.
}
\description{
Function to remove prior imputation.
}
\details{
Observations which have observationFlagVar == imputedFlag will be
    modified to missing. Thus, the value will be updated to NA, the
    observation flag will be updated to missingObservationFlag, and the
    method flag will be updated to missingMethodFlag.
}

