% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNonExistingRecord.R
\name{removeNonExistingRecord}
\alias{removeNonExistingRecord}
\title{Function to remove non-existing record}
\usage{
removeNonExistingRecord(data, areaVar = "geographicAreaM49",
  itemVar = "measuredItemCPC", elementVar = "measuredElement",
  yearVar = "timePointYears", flagObsVar = "flagObservationStatus",
  flagMethodVar = "flagMethod", valueVar = "Value")
}
\arguments{
\item{data}{The data containing non-existing record}

\item{areaVar}{The column name corresponding to the geographic
area.}

\item{itemVar}{The column name corresponding to the commodity
item.}

\item{elementVar}{The column name corresponding to the measured
element.}

\item{yearVar}{The column name corresponding to the year.}

\item{flagObsVar}{The column name corresponding to the observation
status flag.}

\item{flagMethodVar}{The column name corresponding to the method
flag.}

\item{valueVar}{The column name corresponding to the value.}
}
\value{
Data with non-existing records omitted.
}
\description{
Sometimes non-existing records (observation status flag and method flag and
value all NA) are returned by the database, or can be created by
denormalising the data. This function removes these records.
}

