% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSingleEntryCountry.R
\name{removeSingleEntryCountry}
\alias{removeSingleEntryCountry}
\title{This function removes countries that has only one observation}
\usage{
removeSingleEntryCountry(data, params = defaultProcessingParameters())
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProcessingParameters() for a starting
point.}
}
\value{
A data.table where countries with only a single
    observation are omitted from the original data.
}
\description{
TODO (Michael): The by key (param$byKey) should be extended, the
                criteria should be 1 observation per
                country/item/element combination.
}
\details{
NOTE (Michael): The application of the function is questionable. I
                would impute the series even if there is only one
                dataset. It is the best given the information set,
                and we should aim to improve the data collection.
}

