% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoCompare.R
\name{repoCompare}
\alias{repoCompare}
\title{Compare installed packages with those in a remote repository}
\usage{
repoCompare(local_packages = installed.packages(),
  foreign_packages = available.packages())
}
\arguments{
\item{local_packages}{array. Object returned by functions such as 
\code{installed_packages}}

\item{foreign_packages}{array. Object returned by functions such as 
\code{available.packages}}
}
\description{
It's very easy to have one's packages fall out of alignment with the 
Statistical Working System's. This function will help detect where the 
packages you have installed are different to those on the system or in any
other repository.
}
