% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore0M.R
\name{restore0M}
\alias{restore0M}
\title{This function is the reverse of \code{remove0M} which replaces
(NA, M) with (0, M)}
\usage{
restore0M(data, valueVars, flagVars, missingFlag = "M")
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{valueVars}{The value of the observation.}

\item{flagVars}{The flag of the observation.}

\item{missingFlag}{The value of the flag which denotes missing value.}
}
\value{
The data object with all the (0, M) replaced with (NA, M)
}
\description{
This function is the reverse of \code{remove0M} which replaces
(NA, M) with (0, M)
}

