% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setQueryKey.R
\name{setQueryKey}
\alias{setQueryKey}
\title{Set Query Key}
\usage{
setQueryKey(dimension, datasetkey = swsContext.datasets[[1]], newKey)
}
\arguments{
\item{dimension}{character. Name of dimension you wish to set from
DatasetKey}

\item{datasetkey}{Datasetkey. DatasetKey from which you wish to set
dimensions}

\item{newKey}{character. Keys you wish to set}
}
\description{
Set a data key for the query
}
