% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sws_query.r
\name{sws_query}
\alias{sws_query}
\title{Get data from FAO statistical working system (SWS).}
\usage{
sws_query(area, item, ele, year, symb = T, melted = TRUE, value.names = T,
  stringsAsFactors = default.stringsAsFactors(), dbquery,
  class.path = file.path(getwd(), "ojdbc14.jar"), user = "demo",
  pass = "demo")
}
\arguments{
\item{area}{numeric or character vector with countries' ids (numeric) or 
names (character).}

\item{item}{numeric or character vector with items' ids (numeric) or names 
(character).}

\item{ele}{numeric or character vector with elements' ids (numeric) or names 
(character).}

\item{year}{numeric vector with years to fetch.}

\item{symb}{optional character vector with symbols (flags) to fetch data type
of.}

\item{melted}{logical, TRUE by default. Should the result data be returned in
long format (instead of one column per every year two columns: year and 
corresponded value).}

\item{value.names}{logical, TRUE by default. Should identificational vectors 
(area and item) be converted from numeric ids to names.}

\item{stringsAsFactors}{logical. Should character 
identificational vectors be converted to factors.}

\item{dbquery}{optional string with SQL-query to request from SWS.}

\item{class.path}{optional string with path to Oracle Java Data Base
Connectivity library. By default 'ojdbc14.jar' in the working directory.}

\item{user}{optional string with name of SWS DB user instead of default user 
'demo'.}

\item{pass}{optional string with password of SWS DB user instead of default 
  password 'demo'.
  
RJDBC stringr reshape2}
}
\value{
data.frame with results from SWS DB.
}
\description{
Using this function is discouraged and shouldn't be done unless making
comparisons between the old and new working system. This function will not
work on the SWS. This is intentional.
}

