% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toLowerCamel.R
\name{toLowerCamel}
\alias{toLowerCamel}
\title{Function to convert string to lower camel case}
\usage{
toLowerCamel(string)
}
\arguments{
\item{string}{The character string}
}
\description{
Function to convert string to lower camel case
}

