% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDownloadValidate.R
\name{treeDownloadValidate}
\alias{treeDownloadValidate}
\title{Download Commodity Tree and perform validation}
\usage{
treeDownloadValidate(areakeys, yearvals)
}
\arguments{
\item{areakeys}{A character vector containing the M49 country codes}

\item{yearvals}{A character vector containing years to be downloaded}
}
\value{
The Valid Commodity Tree
}
\description{
This function downoad commodity tree form "suafbs" Domain and perform Flag Validation and Value Validation
}
