% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateTreeShares.R
\name{validateTreeShares}
\alias{validateTreeShares}
\title{Validate shares of Commodity Trees

This function perform Flag Validation for the commodity Tree
Validation. 
The possible flags are (decided with Data Team): 

Extraction Rates:
(T,-) = validated up do 2013 - protected
(E,t) = copied from 2014 onwards - not protected but that do not change during standardization process
(E,f) = manually changed by the user - protected

Shares: 
(E,-) = coming from old methodology - NOT protected. These values willbe overwritten
at any run of the module, except the values of oils, which are kept, unless manually changed
(E,f) = manually changed by the user - protected
(I,i) = calculated by the module - not protected}
\usage{
validateTreeShares(tree = NULL)
}
\arguments{
\item{tree}{the commodity tree to check}
}
\value{
a list of two elements: 
messageSH, is a message with the information of validity of flags and 
invalidSH is a data.table which contains invalid Flags or invalid Values or is empty, if everything is valid
}

