% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOutputFlag.R
\name{checkOutputFlags}
\alias{checkOutputFlags}
\title{This is a function to check whether the flags in the data confine
to expectation.}
\usage{
checkOutputFlags(data, flagObservationVar = "flagObservationStatus",
  flagObservationExpected, flagMethodVar = "flagMethod", flagMethodExpected,
  returnData = TRUE, normalised = TRUE,
  denormalisedKey = "measuredElement")
}
\arguments{
\item{data}{The data.table object to be checked}

\item{flagMethodVar}{The column name corresponding to the
method flag.}

\item{flagMethodExpected}{The value of the method flag expected in
the output.}

\item{returnData}{logical, whether the data should be returned}

\item{normalised}{logical, whether the data is normalised}

\item{denormalisedKey}{optional, only required if the input data is not}

\item{flagObservationStatusVar}{The column name corresponding
to the observation status flag.}

\item{flagObservationStatusExpected}{The value of the observation
status flag expected in the output.}
}
\value{
The original data is returned if all the flag matches
    those expected, otherwise an error is raised.
}
\description{
This is a function to check whether the flags in the data confine
to expectation.
}

