% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeYield.R
\name{computeYield}
\alias{computeYield}
\title{Function to compute and update yield}
\usage{
computeYield(productionValue, productionObservationFlag, areaHarvestedValue,
  areaHarvestedObservationFlag, yieldValue, yieldObservationFlag,
  yieldMethodFlag, newMethodFlag, flagTable = faoswsFlag::faoswsFlagTable,
  data, unitConversion = 1)
}
\arguments{
\item{productionValue}{The column name corresponding to production
value.}

\item{productionObservationFlag}{The column name corresponding to
the observation flag of production.}

\item{areaHarvestedValue}{The column name corresponding to area
harvested value.}

\item{areaHarvestedObservationFlag}{The column name corresponding
to the observation flag of area harvested.}

\item{yieldValue}{The columne name corresponding to yield value.}

\item{yieldObservationFlag}{The column name corresponding to the
observation flag of yield.}

\item{yieldMethodFlag}{The column name corresponding to the method flag for
yield.}

\item{newMethodFlag}{The character value that should be placed in the
yieldMethodFlag column when yield is computed.}

\item{flagTable}{see \code{\link[faoswsFlag]{faoswsFlagTable}}.}

\item{data}{The data.table object containing the data.}

\item{unitConversion}{yield is computed as (production)/(area harvested)*
(unit conversion).}
}
\description{
Function to compute and update yield
}

