% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameData.R
\name{nameData}
\alias{nameData}
\title{Name SWS dataset}
\usage{
nameData(domain, dataset, dt, except, append = "_description",
  returnCodes = TRUE)
}
\arguments{
\item{domain}{character. Domain of dataset}

\item{dataset}{character. Name of dataset}

\item{dt}{data.table or data.frame. The table to which you wish to add 
names. Its column names must correspond to the dimensions of the dataset 
specified. Use \code{\link{GetDatasetConfig}} to find the dimension names 
of a dataset}

\item{except}{character. Vector of names in the table that you do not wish 
named. For example, "timePointYears"}

\item{append}{character. Suffix to add to column name of names. If "" and
\code{returnCodes} is FALSE, then adds no suffix to the final names}

\item{returnCodes}{logical. If TRUE, descriptions are returned as columns 
after the codes. If FALSE, the codes are omitted from the output}
}
\description{
Adds human-readable names to keys automatically.
}
\examples{
 
 \dontrun{
 dat = structure(list(geographicAreaM49 = c("276", "36", "380", "4",
"716"), Value = c(0, 0.0555555555555556, 0, 0, 0.2)), row.names = c(NA,
 -5L), class = c("data.table", "data.frame"), .Names = c("geographicAreaM49",
 "Value"), sorted = "geographicAreaM49")
 
 nameData("agriculture", "aproduction", dat, except = "timePointYears")
 }
 
}
