% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{preProcessing}
\alias{preProcessing}
\title{This function performs manipulation of the data that are standard
after the data is retrieved from the data base.}
\usage{
preProcessing(data)
}
\arguments{
\item{data}{The data.table object}
}
\value{
A data.table with standard pre processing steps
    performed.
}
\description{
NOTE (Michael): The data is assumed to be normalised.
}

